requirements.txt
anthropic>=0.40.0
openai>=1.50.0
prompt_toolkit>=3.0.0
pyyaml>=6.0
